Moved subpackages of //hadoop one level up
