+++ hadoop/GraknInputFormat.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.hadoop.cql;
+package grakn.core.hadoop;
 
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.datastax.oss.driver.api.core.cql.ResultSet;
@@ -94,7 +94,7 @@ import static java.util.stream.Collectors.toMap;
  */
 
 public class GraknInputFormat extends org.apache.hadoop.mapreduce.InputFormat<Long, Row> implements org.apache.hadoop.mapred.InputFormat<Long, Row> {
-    public static final String MAPRED_TASK_ID = "mapred.task.id";
+    private static final String MAPRED_TASK_ID = "mapred.task.id";
     private static final Logger LOG = LoggerFactory.getLogger(GraknInputFormat.class);
 
     private String keyspace;
@@ -122,7 +122,7 @@ public class GraknInputFormat extends org.apache.hadoop.mapreduce.InputFormat<Lo
         return new GraknCqlRecordReader();
     }
 
-    protected void validateConfiguration(Configuration conf) {
+    private void validateConfiguration(Configuration conf) {
         if (ConfigHelper.getInputKeyspace(conf) == null || ConfigHelper.getInputColumnFamily(conf) == null) {
             throw new UnsupportedOperationException("you must set the keyspace and table with setInputColumnFamily()");
         }
